moved mixins to the location where they are used
